moved user folder into domain
